+++ docs/src/markdown/about/changelog.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## 0.10a18
+
+- **NEW**: `interpolate` can now accept multiple colors and will return an interpolation function that spans all
+  specified colors via the range of `[0..1]`.
+- **NEW**: `step` can now accept multiple colors and will return colors from the entire range.
+- **NEW**: `compose` can now accept multiple colors and will return a result where all colors are layered on top of each
+  other.
+- **NEW**: `new` method does not need to be a `classmethod`. Make it a normal method on the instance.
+- **FIX**: Actually make `mix` default to `lab` like `interpolate` and friends do.
+
 ## 0.1.0a17
 
 - **FIX**: Ensure that both the Bradford CAT and the XYZ transformation matrix all use ASTM E308-01 white points. This